demo_site/requirements.txt
Django==1.6.1
PyYAML==3.10
dj-static==0.0.5
django-pipeline==1.3.20
django-twitter-bootstrap==3.1.0
futures==2.1.6
gunicorn==18.0
pystache==0.5.3
static==1.0.2